translator_scripts copy/requirements.txt
anthropic>=0.22.0
python-dotenv>=1.0.0
tqdm>=4.66.0
pathlib>=1.0.1
argparse>=1.4.0 